Stop bundling the Codex and Claude Code CLIs in the desktop app
The desktop installer shipped a full copy of both native CLIs: the
@openai/codex platform packages are 310-409 MiB each and the
@anthropic-ai/claude-agent-sdk platform packages 245-262 MiB, and the
macOS universal build carried two of each. That was well over half of a
~1.4 GiB packaged app.

Nothing required them to be there. Both SDKs' JavaScript is inlined into
dist/main by esbuild, and the binary resolvers in codexImport.ts and
claudeAgentImport.ts already probe the global npm prefix and PATH, which
is how the VS Code extension has always worked — it ignores node_modules
entirely. Installing the CLI once per machine now serves every TeXRA
surface, and the Integrations card's "Install in Terminal" action already
drives that flow on the desktop.

Moving both SDKs to devDependencies is what does the work: electron-builder
copies the node_modules of production dependencies only. With the payload
gone, the asarUnpack rules, the mac x64ArchFiles override, and the afterPack
hook that pruned non-matching platforms all become dead config, so this
deletes them along with the payload-budget baselines they were measured
against.

verify-desktop-package.mjs flips from "the expected platform package is
present and within budget" to "no Codex or Claude Code payload is present
at all", covering both the direct and pnpm-store node_modules layouts
inside and outside app.asar, so the binaries cannot creep back in.
