+++ packages/desktop/electron-builder.yml
@@ -13,16 +13,19 @@ protocols:
       - texra
 asar: true
 asarUnpack:
-  - '**/node_modules/@openai/codex/**'
-  - '**/node_modules/@openai/codex-*/**'
   # node-pty loads a native .node binary at runtime, which cannot be read from
   # inside an asar archive.
   - '**/node_modules/node-pty/**'
 npmRebuild: false
-afterPack: ./scripts/prune-desktop-codex-payload.mjs
 directories:
   output: dist-packaged
   buildResources: build
+# electron-builder copies the node_modules of production dependencies only.
+# The Codex and Claude Code SDKs are therefore devDependencies: esbuild inlines
+# their JavaScript into dist/main, while their platform packages — 250-410 MiB
+# of native CLI binary each — stay out of the app, which resolves a
+# user-installed CLI at runtime instead. scripts/verify-desktop-package.mjs
+# fails the build if any of them reappear in the packaged app.
 files:
   - from: dist
     to: dist
@@ -44,7 +47,6 @@ extraResources:
 mac:
   category: public.app-category.productivity
   icon: build/icon.icns
-  x64ArchFiles: '**/app.asar.unpacked/node_modules/{@openai/codex/vendor,@openai/codex-*/vendor,@anthropic-ai/claude-agent-sdk-*}/**'
   target:
     - target: dmg
       arch:
