+++ packages/desktop/package.json
@@ -24,6 +24,8 @@
     "test:e2e": "playwright test"
   },
   "devDependencies": {
+    "@anthropic-ai/claude-agent-sdk": "^0.3.220",
+    "@openai/codex-sdk": "^0.146.0",
     "@playwright/test": "^1.62.0",
     "@types/semver": "^7.7.1",
     "electron": "^43.2.0",
@@ -34,11 +36,9 @@
     "vite": "^8.1.5"
   },
   "dependencies": {
-    "@anthropic-ai/claude-agent-sdk": "^0.3.220",
     "@fontsource-variable/geist": "^5.3.0",
     "@fontsource-variable/jetbrains-mono": "^5.3.0",
     "@fortawesome/free-solid-svg-icons": "^7.3.1",
-    "@openai/codex-sdk": "^0.146.0",
     "@sentry/electron": "^7.16.0",
     "fix-path": "^5.0.0",
     "lucide-static": "^1.27.0",
