--- packages/desktop/scripts/prune-desktop-codex-payload.mjs
@@ -1,1 +1,0 @@
-export { default } from '../../../scripts/prune-desktop-codex-payload.mjs';
